+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.1.0rc1](https://github.com/nikkelarsson/bargets/releases/tag/v0.1.0rc1) -- February 19 2022
+### Added
+- Functionality to configure widgets' looks with config file bargets.yaml
+- Notification to pop up when battery is fully charged
+
+### Changed
+- Don't suspend machine if charger is plugged - even if charge is critically low
+
 ## [0.1.0a0](https://github.com/nikkelarsson/bargets/releases/tag/v0.1.0a0) -- February 05 2022
 ### Added
 - First release of Barget